Refactor object base and references
- Separate generic scene-object behavior from rooted-object behavior.
- Move Object and ObjectReference onto RootedObjectBase.
- Import ObjectType from its dedicated lightweight module.
