+++ isaaclab_arena/assets/object_type.py
@@ -9,10 +9,6 @@ Kept dependency-free so it can be imported by pure-Python spec / schema modules
 (e.g. arena_env_graph_spec.py) without dragging in isaaclab/omni/pxr. This is
 important because importing pxr before SimulationApp starts breaks Kit
 extensions like omni.kit.usd.mdl during pytest collection.
-
-`isaaclab_arena.assets.object_base` re-exports `ObjectType` from here so existing
-`from isaaclab_arena.assets.object_base import ObjectType` consumers keep
-working with a single source of truth.
 """
 
 from enum import Enum
