+++ isaaclab_arena/tests/test_reference_objects.py
@@ -329,8 +329,8 @@ def _test_reference_objects_with_background_pose(background_pose: Pose, tmp_path
 
     from isaaclab.managers import SceneEntityCfg
 
-    from isaaclab_arena.assets.object_base import ObjectType
     from isaaclab_arena.assets.object_reference import ObjectReference, OpenableObjectReference
+    from isaaclab_arena.assets.object_type import ObjectType
     from isaaclab_arena.cli.isaaclab_arena_cli import arena_env_builder_cfg_from_argparse, get_isaaclab_arena_cli_parser
     from isaaclab_arena.embodiments.franka.franka import FrankaIKEmbodiment
     from isaaclab_arena.environments.arena_env_builder import ArenaEnvBuilder
