+++ isaaclab_arena_environments/droid_table_multi_object_placement_environment.py
@@ -84,8 +84,8 @@ class DroidTableMultiObjectPlacementEnvironment(ArenaEnvironmentFactory[DroidTab
         """Build the environment from its typed configuration."""
         from isaaclab.envs.common import ViewerCfg
 
-        from isaaclab_arena.assets.object_base import ObjectType
         from isaaclab_arena.assets.object_reference import ObjectReference
+        from isaaclab_arena.assets.object_type import ObjectType
         from isaaclab_arena.environments.isaaclab_arena_environment import IsaacLabArenaEnvironment
         from isaaclab_arena.relations.relations import IsAnchor
         from isaaclab_arena.scene.scene import Scene
